+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/General.kt
@@ -3,8 +3,7 @@
 package csense.kotlin.tests.assertions
 
 import csense.kotlin.annotations.numbers.*
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.contract
+import kotlin.contracts.*
 import kotlin.test.*
 
 /**
@@ -226,8 +225,14 @@ public fun <@kotlin.internal.OnlyInputTypes T> T.assertByEquals(
     )
 }
 
+@Suppress("RedundantNullableReturnType", "NOTHING_TO_INLINE")
+public inline fun <T> (T & Any).nullable(): T? {
+    return this
+}
+
 
 public object GeneralStrings {
     public const val assertCalledMessage: String = "Should be called, but did not get called enough times"
     public const val assertNotCalledMessage: String = "Should not be called but got called anyway"
 }
+