+++ src/commonTest/kotlin/csense/kotlin/tests/assertions/ArraysTest.kt
@@ -1206,7 +1206,7 @@ class ArraysTest {
     }
 
 
-    class ArrayTAssertExpected {
+    class ArrayOutTAssertExpected {
 
         @Test
         fun emptyToEmpty() {