+++ src/commonTest/kotlin/csense/kotlin/tests/assertions/PairTest.kt
@@ -8,13 +8,13 @@ class PairTest {
         @Test
         fun nonNullShouldWork() {
             val p1 = Pair("a", "1")
-            p1.assert("a" to "1")
+            p1.assertByEquals("a" to "1")
         }
 
         @Test
         fun nullToNonNullShouldThrow() = assertThrows<Throwable> {
             val p1: Pair<String, String>? = null
-            p1.assert("a" to "1")
+            p1.assertByEquals("a" to "1")
         }
     }
 }